#2455 Split LFRicDofmaps from dynamo0p3
