+++ src/psyclone/domain/lfric/__init__.py
@@ -71,6 +71,7 @@ from psyclone.domain.lfric.lfric_invokes import LFRicInvokes
 from psyclone.domain.lfric.lfric_scalar_args import LFRicScalarArgs
 from psyclone.domain.lfric.lfric_loop_bounds import LFRicLoopBounds
 from psyclone.domain.lfric.lfric_kern_metadata import LFRicKernMetadata
+from psyclone.domain.lfric.lfric_dofmaps import LFRicDofmaps
 
 
 __all__ = [
@@ -83,6 +84,7 @@ __all__ = [
     'LFRicArgDescriptor',
     'LFRicCollection',
     'LFRicConstants',
+    'LFRicDofmaps',
     'LFRicExtractDriverCreator',
     'LFRicFields',
     'LFRicInvoke',
